+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Basic step list validator
 - Final steps list printing via window button
 - Slightly faster project scrubbing
+- Readme.md generator
 
 # 1.0.0